skills/mem0-memory/scripts/requirements.txt
# Mem0 Python SDK
# Install with: pip install mem0ai
mem0ai>=1.0.0
